requirements.txt
streamlit
PyPDF2
textstat
Pillow
PyMuPDF